feat(media): implement lifecycle reconciliation and playback payload building
- Added lifecycle reconciler to manage media object states and usage.
- Introduced playback payload builder for constructing playback data.
- Refactored routes to utilize new lifecycle and playback functionalities.
- Enhanced notification broadcasting with internal audit logging.
- Created native backend catalog for processing various tasks including billing and notifications.
- Implemented runtime bootstrap for managing service dependencies.
- Added tests for native billing and notifications to ensure functionality.
- Updated processor catalog to support backend mode selection (http/native).

+++ services/openvibe-billing/server/db.js
@@ -207,7 +207,7 @@ const SCHEMA_SQL = `
         CREATE INDEX IF NOT EXISTS idx_billing_audit_actor ON billing_audit(actor_type, actor_id);
 
         CREATE TABLE IF NOT EXISTS billing_economy_state (
-            id          INTEGER PRIMARY KEY CHECK (id = 1),
+            id          BIGINT PRIMARY KEY CHECK (id = 1),
             frozen      INTEGER NOT NULL DEFAULT 0,
             reason      TEXT,
             updated_at  DATETIME DEFAULT CURRENT_TIMESTAMP,
